Add files via upload
Faculty takes lot of time to adjust their classes when they are on leave. They have to check
manually timetables of all other faculty to know whether they are free or contact each and every
faculty and ask them if they are free. Lot of time is wasted in this process. To overcome this
problem, we designed a website in which we upload timetables of all faculties and using an
algorithm we display the names and phone numbers of free faculties of selected slot. This solves
the problem in couple of minutes.  I designed a website to solve the above problem. In the website, faculty registers with his/her
username, phone number and other details. He/She logins to the website and checks for the free
faculty available in those particular slots.